refactor: moved to data file
