+++ tests/unit/test_data.py
@@ -24,11 +24,88 @@ from unittest.mock import MagicMock
 
 import jax.numpy as jnp
 import numpy as np
+import pytest
 from jax import Array
 from jax.typing import ArrayLike
 
 import coreax.data
 import coreax.reduction
+
+
+class TestData:
+    """Test operation of Data class."""
+
+    def test_uniform_weights(self):
+        """Test that if no weights are given a uniform weight vector is made."""
+        original_data = jnp.array([1, 2, 3])
+        n = original_data.shape[0]
+
+        data = coreax.data.Data(data=original_data, weights=None)
+        np.testing.assert_array_almost_equal(
+            data.weights, jnp.broadcast_to(1 / n, (n,)), decimal=5
+        )
+
+    def test_invalid_data_and_weight_dimensions(self):
+        """
+        Test that __check_init__ raises expected errors.
+        """
+        original_data = jnp.array([1, 2, 3])
+        with pytest.raises(
+            ValueError, match="Leading dimensions of 'weights' and 'data' must be equal"
+        ):
+            coreax.data.Data(
+                data=original_data, weights=jnp.ones(original_data.shape[0] + 1)
+            )
+
+
+class TestSupervisedData:
+    """Test operation of SupervisedData class."""
+
+    def test_uniform_weights(self):
+        """Test that if no weights are given a uniform weight vector is made."""
+        original_data = jnp.array([1, 2, 3])
+        original_supervision = jnp.array([4, 5, 6])
+        n = original_data.shape[0]
+
+        data = coreax.data.SupervisedData(
+            data=original_data, supervision=original_supervision, weights=None
+        )
+        np.testing.assert_array_almost_equal(
+            data.weights, jnp.broadcast_to(1 / n, (n,)), decimal=5
+        )
+
+    def test_invalid_data_and_weight_dimensions(self):
+        """
+        Test that __check_init__ raises ValueError when weights array has bad dimension.
+        """
+        original_data = jnp.array([1, 2, 3])
+        original_supervision = jnp.array([4, 5, 6])
+
+        with pytest.raises(
+            ValueError, match="Leading dimensions of 'weights' and 'data' must be equal"
+        ):
+            coreax.data.SupervisedData(
+                data=original_data,
+                supervision=original_supervision,
+                weights=jnp.ones(original_data.shape[0] + 1),
+            )
+
+    def test_invalid_data_and_supervision_dimensions(self):
+        """
+        Test that __check_init__ raises ValueError when supervision has bad dimension.
+        """
+        original_data = jnp.array([1, 2, 3])
+        original_supervision = jnp.array([4, 5])
+
+        with pytest.raises(
+            ValueError,
+            match="Leading dimensions of 'supervision' and 'data' must be equal",
+        ):
+            coreax.data.SupervisedData(
+                data=original_data,
+                supervision=original_supervision,
+                weights=jnp.ones(original_data.shape[0] + 1),
+            )
 
 
 class DataReaderConcrete(coreax.data.DataReader):
